Update smtp.php
Error 250 can contain anything. For example: 550 Message was not accepted -- invalid mailbox
Why does this error cause all of the store's functions to malfunction? This is a minor error that should not affect order processing in any way. 
I suggest that these errors should be removed as fatal ones and just write them into error log.