we trust chance - Update file.php
If we trust chance, but we have to trust it to the end.

Otherwise, our glob function is constantly idling. 

There will always be some files in the cache, and it takes time and memory to store an array of files.

Now the cache will be read and cleared only if an accident has occurred.